Some native changes
