+++ android/lib/src/main/java/com/swmansion/gesturehandler/OnTouchEventListener.java
@@ -7,4 +7,5 @@ import com.swmansion.gesturehandler.GestureHandler;
 public interface OnTouchEventListener<T extends GestureHandler> {
   void onTouchEvent(T handler, MotionEvent event);
   void onStateChange(T handler, int newState, int oldState);
+  void onPassBounds(T handler, boolean isOutside);
 }
